pyproject.toml
[ build-system ]
requires = [ "setuptools >= 64.0" ]
build-backend = "setuptools.build_meta"


[ project ]
name = "sharelock"
version = "0.1.0"
description = "Create encrypted files whose decryption key is split and shared between multiple individuals."
dependencies = [
    "pycryptodome",
    "eciespy",
    "eth-keys",
    "coincurve",
]
requires-python = ">= 3.8"
readme = "README.md"
license = "MIT"
license-files = [ "LICENSE" ]


[project.scripts]
sharelock = "sharelock.sharelock:main"

